+++ src/bootstrap/bootstrap.ts
@@ -1,13 +1,17 @@
 import {
+  ModuleLoadingError,
   WorkspaceContextError,
-  isModuleLocalPath,
   loadWorkspaceConfiguration,
 } from '@causa/workspace';
-import { exec } from 'child_process';
-import { mkdir, rm, stat, writeFile } from 'fs/promises';
+import {
+  CAUSA_FOLDER,
+  setUpCausaFolder,
+} from '@causa/workspace/initialization';
+import { stat } from 'fs/promises';
 import { join, resolve } from 'path';
 import { Logger } from 'pino';
 import { fileURLToPath } from 'url';
+import { RunCliOptions } from '../cli.js';
 import { showHelpForCommand } from '../command-help.js';
 import {
   GlobalCliOptions,
@@ -15,18 +19,42 @@ import {
   parseGlobalOptions,
 } from '../command.js';
 import { createLogger } from '../logger.js';
-import { ModuleInstallationError } from './errors.js';
 import { runCliInWorkerThread } from './worker.js';
-
-/**
- * The folder placed at the root of Causa workspaces, containing the npm installation.
- */
-const CAUSA_FOLDER = '.causa';
 
 /**
  * The expected location of the CLI script (which can be called as a worker thread), relative to the Causa folder.
  */
 const CAUSA_CLI_LOCATION = 'node_modules/@causa/cli/dist/bootstrap/cli.js';
+
+/**
+ * The result of {@link setUpCausaFolderIfNeeded}, which contains the detected workspace and Causa folder location.
+ */
+type CausaFolderSetupResult = {
+  /**
+   * The location of the workspace root.
+   */
+  rootPath: string;
+
+  /**
+   * The location of the Causa CLI script.
+   */
+  causaCliLocation: string;
+
+  /**
+   * The modules to install in the Causa folder, loaded from the configuration.
+   */
+  modules: Record<string, string>;
+
+  /**
+   * Whether the current process is running from the detected workspace installation.
+   */
+  isRunningFromWorkspace: boolean;
+
+  /**
+   * Whether the Causa folder was initialized and modules were installed during the setup.
+   */
+  didSetUp: boolean;
+};
 
 /**
  * The entrypoint of the CLI.
@@ -39,7 +67,7 @@ const CAUSA_CLI_LOCATION = 'node_modules/@causa/cli/dist/bootstrap/cli.js';
  */
 export async function bootstrapCli(args: string[]): Promise<number> {
   let cliOptions: GlobalCliOptions;
-  let causaCliLocation: string | null;
+  let setupResult: CausaFolderSetupResult;
   let logger: Logger | undefined;
   try {
     cliOptions = parseGlobalOptions(args);
@@ -49,27 +77,84 @@ export async function bootstrapCli(args: string[]): Promise<number> {
 
     logger = createLogger({ verbose: cliOptions.verbose });
 
-    causaCliLocation = await setUpCausaFolderIfNeeded(workingDirectory, logger);
+    setupResult = await setUpCausaFolderIfNeeded(workingDirectory, logger);
   } catch (error: any) {
-    const errorLogger = logger ?? createLogger();
+    logger = logger ?? createLogger();
 
     const message = error.message ?? error;
-    errorLogger.error(`❌ ${message}`);
+    logger.error(`❌ ${message}`);
 
     if (error instanceof WorkspaceContextError) {
       // This can be due to the CLI being run outside of a Causa workspace, in which case the user might need help.
-      showHelpForCommand(createBaseCommand(), errorLogger);
+      showHelpForCommand(createBaseCommand(), logger);
     }
 
     return 1;
   }
 
-  if (causaCliLocation === null) {
-    const { runCli } = await import('../cli.js');
-    return await runCli(args, cliOptions);
-  }
+  return await runCliFromSetupResult(args, setupResult, cliOptions, logger);
+}
 
-  return await runCliInWorkerThread(causaCliLocation, args, cliOptions);
+/**
+ * Runs the CLI, using the given setup result to determine how to run it.
+ * If the current process is running from the workspace installation, it will run the CLI directly. Otherwise, it is run
+ * as a worker thread. Also, if modules have not been installed during the setup, module version errors will be caught
+ * and will trigger the setup again, forcing the modules to be installed.
+ *
+ * @param args The arguments to pass to the CLI, without the node executable and the script name.
+ * @param setupResult The result of the setup operation, used to determine whether to run the CLI directly or as a
+ *   worker thread.
+ * @param globalCliOptions The options to pass to the CLI.
+ * @param logger The logger to use.
+ * @returns The exit code of the CLI.
+ */
+async function runCliFromSetupResult(
+  args: string[],
+  setupResult: CausaFolderSetupResult,
+  globalCliOptions: GlobalCliOptions,
+  logger: Logger,
+): Promise<number> {
+  const runCliOptions: RunCliOptions = {
+    ...globalCliOptions,
+    rethrowModuleLoadingError: !setupResult.didSetUp,
+  };
+
+  try {
+    if (setupResult.isRunningFromWorkspace) {
+      const { runCli } = await import('../cli.js');
+      return await runCli(args, runCliOptions);
+    }
+
+    return await runCliInWorkerThread(
+      setupResult.causaCliLocation,
+      args,
+      runCliOptions,
+    );
+  } catch (error: any) {
+    // Not checking the actual type of the error, as it might come from a different version of the workspace module.
+    if (
+      !(error as ModuleLoadingError).requiresModuleInstall ||
+      !runCliOptions.rethrowModuleLoadingError
+    ) {
+      throw error;
+    }
+
+    logger.warn(
+      '📦 Module version changes detected in the configuration, reinstalling modules in the Causa folder.',
+    );
+
+    await setUpCausaFolderWithCli(
+      setupResult.rootPath,
+      setupResult.modules,
+      logger,
+    );
+
+    return await runCliInWorkerThread(
+      setupResult.causaCliLocation,
+      args,
+      globalCliOptions,
+    );
+  }
 }
 
 /**
@@ -78,12 +163,12 @@ export async function bootstrapCli(args: string[]): Promise<number> {
  *
  * @param workingDirectory The working directory from which to look for the workspace.
  * @param logger The logger to use.
- * @returns The location of the CLI script, or `null` if the current process is running from the workspace installation.
+ * @returns A {@link CausaFolderSetupResult} object.
  */
 async function setUpCausaFolderIfNeeded(
   workingDirectory: string,
   logger: Logger,
-): Promise<string | null> {
+): Promise<CausaFolderSetupResult> {
   logger.debug(
     `📂 Looking for workspace to bootstrap starting from '${workingDirectory}'.`,
   );
@@ -95,16 +180,17 @@ async function setUpCausaFolderIfNeeded(
   );
 
   const causaFolder = join(rootPath, CAUSA_FOLDER);
+  const causaCliLocation = join(causaFolder, CAUSA_CLI_LOCATION);
+  const modules = configuration.get('causa.modules') ?? {};
+  const partialResult = { causaCliLocation, rootPath, modules };
 
   const sourceFilePath = fileURLToPath(import.meta.url);
   if (sourceFilePath.startsWith(causaFolder)) {
     logger.debug(
       `📂 Already running from the workspace installation in '${causaFolder}'.`,
     );
-    return null;
+    return { ...partialResult, isRunningFromWorkspace: true, didSetUp: false };
   }
-
-  const causaCliLocation = join(causaFolder, CAUSA_CLI_LOCATION);
 
   try {
     await stat(causaCliLocation);
@@ -112,7 +198,7 @@ async function setUpCausaFolderIfNeeded(
     logger.debug(
       `📂 Found existing workspace installation in '${causaFolder}'.`,
     );
-    return causaCliLocation;
+    return { ...partialResult, isRunningFromWorkspace: false, didSetUp: false };
   } catch (error: any) {
     if (error.code !== 'ENOENT') {
       throw error;
@@ -122,89 +208,28 @@ async function setUpCausaFolderIfNeeded(
   logger.info(
     `🎉 Could not find a workspace installation with CLI in '${causaFolder}', creating it.`,
   );
-  const modules = configuration.get('causa.modules') ?? {};
-  await setUpCausaFolder(causaFolder, modules, logger);
+  await setUpCausaFolderWithCli(rootPath, modules, logger);
 
-  return causaCliLocation;
+  return { ...partialResult, isRunningFromWorkspace: false, didSetUp: true };
 }
 
 /**
  * Initializes the Causa folder at the given location, installing the npm modules in the process.
+ * If not specified, the Causa CLI package will be added to the modules to install.
  *
- * @param causaFolder The location of the Causa folder to initialize.
+ * @param rootPath The path to the workspace root.
  * @param modules The modules to install in the Causa folder.
  * @param logger The logger to use.
  */
-async function setUpCausaFolder(
-  causaFolder: string,
+async function setUpCausaFolderWithCli(
+  rootPath: string,
   modules: Record<string, string>,
   logger: Logger,
 ): Promise<void> {
-  await mkdir(causaFolder, { recursive: true });
-
-  await makePackageFile(causaFolder, modules);
-
-  await installModules(causaFolder, logger);
-}
-
-/**
- * Creates the `package.json` file in the Causa folder, containing the given modules.
- * Modules that are local paths will be ignored. If not present, the `@causa/cli` module will be added.
- *
- * @param causaFolder The location of the Causa folder in which to create the package file.
- * @param modules The modules to include in the package file.
- */
-async function makePackageFile(
-  causaFolder: string,
-  modules: Record<string, string>,
-): Promise<void> {
-  const dependencies = { ...modules };
-
-  Object.keys(dependencies).forEach((moduleName) => {
-    if (isModuleLocalPath(moduleName)) {
-      delete dependencies[moduleName];
-    }
-  });
-
-  if (dependencies['@causa/cli'] === undefined) {
-    dependencies['@causa/cli'] = '*';
+  modules = { ...modules };
+  if (modules['@causa/cli'] === undefined) {
+    modules['@causa/cli'] = '*';
   }
 
-  const packagePath = join(causaFolder, 'package.json');
-  await writeFile(packagePath, JSON.stringify({ dependencies }));
-}
-
-/**
- * Installs the npm modules in the Causa folder, after removing the existing `node_modules` folder.
- *
- * @param causaFolder The location of the Causa folder in which to install the modules.
- * @param logger The logger to use.
- */
-async function installModules(
-  causaFolder: string,
-  logger: Logger,
-): Promise<void> {
-  const nodeModulesFolder = join(causaFolder, 'node_modules');
-
-  logger.debug(
-    `🔥 Removing existing node modules folder '${nodeModulesFolder}'.`,
-  );
-  await rm(nodeModulesFolder, { recursive: true, force: true });
-  await rm(join(causaFolder, 'package-lock.json'), { force: true });
-
-  logger.debug(`➕ Installing node modules in '${causaFolder}'.`);
-
-  await new Promise<void>((resolve, reject) => {
-    const child = exec('npm install --quiet', {
-      cwd: causaFolder,
-    });
-    child.stderr?.pipe(process.stderr);
-    child.on('close', (code) => {
-      if (code === 0) {
-        resolve();
-      } else {
-        reject(new ModuleInstallationError());
-      }
-    });
-  });
+  await setUpCausaFolder(rootPath, modules, logger);
 }