+++ src/cli.spec.ts
@@ -1,4 +1,4 @@
-import { WorkspaceContext } from '@causa/workspace';
+import { ModuleLoadingError, WorkspaceContext } from '@causa/workspace';
 import { jest } from '@jest/globals';
 import { mkdtemp, rm } from 'fs/promises';
 import 'jest-extended';
@@ -25,8 +25,10 @@ describe('command', () => {
         environments: { dev: { name: '🪛' } },
         causa: {
           modules: {
+            // Sneaky trick to import a local file as a module (which is not a valid package).
+            // The version should be a local path such that it is not checked.
             [fileURLToPath(new URL('./cli.module.test.ts', import.meta.url))]:
-              '',
+              'file:/some/path',
           },
         },
       });
@@ -187,6 +189,20 @@ describe('command', () => {
       expect(logger.error).toHaveBeenCalledWith(
         expect.stringContaining('Usage:'),
       );
+    });
+
+    it('should rethrow a ModuleLoadingError when requested', async () => {
+      await writeConfiguration(tmpDir, 'causa.yaml', {
+        workspace: { name: 'my-workspace' },
+        causa: { modules: { pino: '0.0.1' } },
+      });
+
+      const actualPromise = runCli(['myFunction'], {
+        workingDirectory: tmpDir,
+        rethrowModuleLoadingError: true,
+      });
+
+      await expect(actualPromise).rejects.toThrow(ModuleLoadingError);
     });
   });
 });